+++ test/acceptance/export/svg.js
@@ -150,7 +150,8 @@ test('SVG format with "the_geom" in skipfields', function(done){
         method: 'GET'
     },{ }, function(res){
         assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
-
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
         assert.deepEqual(JSON.parse(res.body), {
           error:['column "the_geom" does not exist']
         });